+++ packages/reactabular-table/src/index.js
@@ -4,4 +4,5 @@ export { default as Body } from './body';
 export { default as BodyRow } from './body-row';
 export { default as evaluateFormatters } from './evaluate-formatters';
 export { default as evaluateTransforms } from './evaluate-transforms';
+export { default as mergeProps } from './merge-props';